dont consider canonical link when caching
The program has already handled the case when canonical link is not
populated. But it turns out that some websites, instead of setting
canonical correctly or not setting it at all, set it to a URL that is
different from the final URL after all redirections. This means we
cannot know whether to use the canonical link in all cases.

Therefore, do not consider canonical at all during caching. Instead,
just inspect it and report any odd findings.