fix fetch --csv crashing and refactor
This fixes a bug where the fetch --csv command crashed because, unlike
the update command, it was not skipping the csv header row. This in
turn caused the code afterwards to parse string "url" as a URL, which
failed.

This change also fixes the origin of the bug, where the fetch and update
commands used different code to read the beginning of the file. Now both
commands reuse the same code that

- checks that the csv file is not empty
- verifies that the csv header is exactly as expected by the program
- skips the header to advance the reader to the first item

A side benefit is that csv-specific code for checking the header and
writing it back is removed from the update_csv function.