refactor logging contexts
This removes the ad-hoc boilerplate to add context to log entries.

The benefits are less noise, better readability, less room for error,
and cleaner function signatures.

As a side effect it also removes the confusing logging of "(0)"
prefix when the fetch command was invoked with a URL argument. Since
there is no csv file in that context, csv record index was confusing.